Phase 98d: Fix 4 build errors + 10 warnings in clash detection module
All issues are pre-existing in the Clash subsystem (committed April 18
in the clash 1.x–5.x sequence); they were masked by the earlier NU1605
RBush downgrade error that blocked the C# compile.

Errors (4):

CS1739 AabbSweep.cs — RBush 4.0 renamed Envelope constructor params
  from minX/minY/maxX/maxY to MinX/MinY/MaxX/MaxY. Switched to
  positional args so the call survives either naming convention.

CS0103 x2 ExporterIFCUtils not found (ClashSession + ClashExportContext)
  ExporterIFCUtils lives in RevitAPIIFC.dll which isn't referenced by
  StingTools.csproj. Replaced the CreateSubElementGUID call with a
  direct Element.UniqueId read — same string Revit feeds into its IFC
  export pipeline internally, so the clash-kernel's per-element
  identity hash stays equivalent.

CS1061 LinkNode.GetSymbolId not found (ClashExportContext.cs)
  Revit 2025 removed LinkNode.GetSymbolId. Replaced with a stable
  32-bit hash of the linked document's PathName so the clash identity
  keys stay unique across sessions. Also replaced silent-swallow
  catch with StingLog.Warn for diagnostic visibility.

Warnings (10):

CS0618 x6 ElementId.IntegerValue / ElementId(int) obsolete
  Revit 2024 deprecated ElementId.IntegerValue in favour of Value
  (Int64) and the ElementId(int) ctor in favour of ElementId(Int64).
  Updated LiveClashUpdater (3 sites), LiveClashFlag (1 site),
  ClashSession (2 sites), and ClashExportContext (1 site). Other
  callers in BIMManager/ExcelLink/Revision/etc. weren't flagged
  because they already pass long args and resolve to the long
  overload — untouched.

CS0067 x2 ClashSession.OnElementFlagChanged + ClashRowViewModel.PropertyChanged
  Both are part of public API contracts (dock-panel live-clash
  observer subscribes to OnElementFlagChanged; DataGrid two-way
  binding relies on PropertyChanged). Compiler can't see the
  subscribe paths. Suppressed with #pragma warning disable CS0067
  around the declaration — keeps the contract intact instead of
  deleting.

CS0169 BcfSnapshotter._tempView never used
  Field reserved for future session-scoped view caching (avoids
  create/destroy view per snapshot). Suppressed with #pragma warning
  disable CS0169 with an explanatory comment so the lifecycle
  refactor doesn't need to reintroduce it.

+++ StingTools/Clash/LiveClashFlag.cs
@@ -34,7 +34,8 @@ namespace StingTools.Core.Clash
 
         private static void SetParam(Document doc, int elementId, bool value)
         {
-            var el = doc.GetElement(new ElementId(elementId));
+            // ElementId(int) ctor is obsolete in Revit 2024+; use Int64 overload.
+            var el = doc.GetElement(new ElementId((long)elementId));
             if (el == null) return;
             var p = el.LookupParameter(ParamName);
             if (p == null || p.IsReadOnly) return;
